# pkg/docs/swaggers/mead_api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "mead_api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "MeadAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/statistics": {
      "post": {
        "operationId": "MeadAPI_ListStatistics",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mead_apiListStatisticsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "MeadAPI"
        ]
      }
    },
    "/api/user": {
      "post": {
        "operationId": "MeadAPI_CreateUser",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mead_apiCreateUserResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/mead_apiCreateUserRequest"
            }
          }
        ],
        "tags": [
          "MeadAPI"
        ]
      }
    },
    "/api/version": {
      "get": {
        "operationId": "MeadAPI_Version",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/mead_apiVersionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "MeadAPI"
        ]
      }
    }
  },
  "definitions": {
    "RequestManyUsers": {
      "type": "object",
      "properties": {
        "usernames": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "mead_apiCreateUserRequest": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "usernames": {
          "$ref": "#/definitions/RequestManyUsers"
        }
      }
    },
    "mead_apiCreateUserResponse": {
      "type": "object"
    },
    "mead_apiListStatisticsResponse": {
      "type": "object",
      "properties": {
        "stas": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/mead_apiUserStatistic"
          }
        }
      }
    },
    "mead_apiUserStatistic": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "lastConnect": {
          "type": "string"
        },
        "bytesPassed": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "mead_apiVersionResponse": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        },
        "clientTimestamp": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}